Create input fields and save the data provided to those input fields to a list
This initial implementation has three simple input widgets, one of which is a datepicker. The other two input fields are for 'distance traveled' and 'note' respectively. The datepicker allows to pick the date for when the traveling was done. The distance traveled takes only numbers for the users to input the distance they've traveled. The note is for any additional data provided by the user. The 'Save Travel' button saves the data from those fields to a custom list of objects.