Display data at the beginning.
The data that is stored in the device is displayed when opening the app. The screen that displays the data also has a floating action button that lets the user navigate to a screen where they can add data. When the data is saved they are shown a prompt and navigated back to the previous screen while showing the recently added data. This is made possible because of the provider package used for state management throughout the app.