Multiple updates to the code
A details screen for the travel details is created to show more details about the travel. This page will also hold the feature to edit or delete the travel detail.
The additional details provided in the parameters of travel card has been changed because the 'entry' data provided was enough.
The entry provider now has an additional function to sort the entries according to dates which is displayed in the first screen.
Finally, the date picker now only allows one to select date from their current and past time instead of letting them select any date.