Make the stacked card and colors reusable
Previously the stacked cards and the colors for those were created out of the fly when needed. These widgets and color picker will be used widely in the future widgets. To do so the widget itself is now more modular and reusable where needed and have the ability to be tweaked (just the stacked card widget).