Create custom input field and customize add entry screen
The input field when adding travel entries has been made reusable and customized. Validators have been added to this custom input field if validation is required for some inputs. The button in add screen also now is themed to fit the app. Other minor changes include not displaying note if none is added and the flushbar.