Make the flushbar modular and more accessible
Previously, the flushbar that showed user's of a message about any action they completed was generated where needed manually. To remove this redundancy in code, the flushbar can now be shown using a helper function available inside the utils called showFlushbar(). This function has been modified and is reusable throughout the app by passing a message and context.

There are two additional parameters provided which have default value but can be changed when calling the function if needed.